% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_phylogenies.R
\name{check_phylogenies}
\alias{check_phylogenies}
\title{Check if the input is a valid collection of one or more phylogenies}
\usage{
check_phylogenies(phylogenies)
}
\description{
Check if the input is a valid collection of one or more phylogenies
}
