% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_step_type.R
\name{check_step_type}
\alias{check_step_type}
\title{Check if the step type is valid}
\usage{
check_step_type(step_type)
}
\arguments{
\item{step_type}{when between two points, where the second point
 has both a higher x and y coordinat, which y coordinat to follow.
'step_type' can be:
\itemize{
   \item lower maintain the y-coordinat of the leftmost point
   \item upper already use the y-coordinat of the rightmost point
}}
}
\description{
Will \link{stop} if not
}
