% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_average_nltt.R
\name{get_average_nltt}
\alias{get_average_nltt}
\title{Get the average nLTT from a collection of phylogenies}
\usage{
get_average_nltt(phylogenies, dt = 0.001, plot_nltts = FALSE,
  xlab = "Normalized Time", ylab = "Normalized Lineages", replot = FALSE,
  ...)
}
\arguments{
\item{phylogenies}{the phylogenies, where the phylogenies are of type 'phylo'}

\item{dt}{The timestep resolution, where 1/dt is the number of points evaluated}

\item{plot_nltts}{Also plot each nLLT line}

\item{xlab}{Label on the x axis}

\item{ylab}{Label on the y axis}

\item{replot}{If false, start a clean plot. If true, plot the new data over the current}

\item{...}{Plotting options}
}
\value{
Nothing
}
\description{
Get the average nLTT from a collection of phylogenies
}
\examples{
  get_average_nltt(c(ape::rcoal(10), ape::rcoal(10)))
  get_average_nltt(c(ape::rcoal(10), ape::rcoal(20)), dt = 0.1)

}
\author{
Richel Bilderbeek
}

