% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_average_nltt_matrix.R
\name{get_average_nltt_matrix}
\alias{get_average_nltt_matrix}
\title{Get the average nLTT from a collection of phylogenies}
\usage{
get_average_nltt_matrix(phylogenies, dt = 0.001)
}
\arguments{
\item{phylogenies}{the phylogenies, supplied as either a list or a multiPhylo object, where the phylogenies are of type 'phylo'}

\item{dt}{The timestep resolution, where 1/dt is the number of points evaluated}
}
\value{
A matrix of timepoints with the average number of (normalized) lineages through (normalized) time
}
\description{
Get the average nLTT from a collection of phylogenies
}
\examples{
  get_average_nltt_matrix(c(ape::rcoal(10), ape::rcoal(20)))

}
\author{
Richel Bilderbeek
}

