% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_average_nltt_matrix.R
\name{get_average_nltt_matrix_impl_2}
\alias{get_average_nltt_matrix_impl_2}
\title{Internal function. Second implementation of get_average_nltt_matrix. A stub now.}
\usage{
get_average_nltt_matrix_impl_2(phylogenies, dt)
}
\arguments{
\item{phylogenies}{the phylogenies, supplied as either a list or a multiPhylo object, where the phylogenies are of type 'phylo'}

\item{dt}{The timestep resolution, where 1/dt is the number of points evaluated}
}
\value{
A matrix of timepoints with the average number of (normalized) lineages through (normalized) time
}
\description{
Internal function. Second implementation of get_average_nltt_matrix. A stub now.
}
\author{
Richel Bilderbeek
}

