% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_braching_times.R
\name{get_branching_times}
\alias{get_branching_times}
\title{Collect the branching times from the stem age}
\usage{
get_branching_times(phylogeny)
}
\arguments{
\item{phylogeny}{a phylogeny of class 'phylo'}
}
\value{
branching times, in time units before the present
}
\description{
Collect the branching times from the stem age
}
\examples{
  phylogeny <- ape::read.tree(text = "((a:2,b:2):1,c:3);")
  phylogeny$root.edge <- 2
  testthat::expect_true(
    all(nLTT::get_branching_times(phylogeny) == c(5, 3, 2)))
}
\author{
Richel Bilderbeek
}

