% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phylogeny_nltt_matrix.R
\name{get_phylogeny_nltt_matrix}
\alias{get_phylogeny_nltt_matrix}
\title{Extract the nLTT matrix from a phylogeny}
\usage{
get_phylogeny_nltt_matrix(phylogeny)
}
\arguments{
\item{phylogeny}{A phylogeny of type phylo}
}
\value{
a matrix
}
\description{
Extract the nLTT matrix from a phylogeny
}
\author{
Richel Bilderbeek
}

