\name{nLTT-package}
\alias{nLTT-package}
\alias{nLTT}
\docType{package}
\title{
	Package providing functions to visualize the normalized Lineage-Through-Time statistic, and calculate the difference between two nLTT curves.
}
\description{
	This package provides a function to visualize the normalized Lineage-Through-Time (nLTT) statistic, where the number of lineages relative to the maximum number of lineages in a phylogenetic tree is plotted against the relative time between the most common recent ancestor and the present. Furthermore the package provides two functions to calculate either the absolute or the squared difference between two nLTT curves.
}
\details{
\tabular{ll}{
Package: \tab nLTT\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-09-17\cr
License: GPL 2.0
}

}

\references{
 T. Janzen, S. Hoehna, R.S. Etienne, 2015, Approximate Bayesian Computation of diversification rates from molecular phylogenies: introducing a new efficient summary statistic, the nLTT, Methods in Ecology and Evolution, in prep
}

\author{
Thijs Janzen

Maintainer: Thijs Janzen <thijsjanzen@gmail.com>
}
