% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nltt_diff_exact_extinct.R
\name{nltt_diff_exact_calc_extinct}
\alias{nltt_diff_exact_calc_extinct}
\title{Calculates the exact difference between the nLTT
curves of the event times. This includes extinction events.}
\usage{
nltt_diff_exact_calc_extinct(
  event_times,
  species_number,
  event_times2,
  species_number2,
  distance_method
)
}
\arguments{
\item{event_times}{event times of the first phylogeny}

\item{species_number}{the number of species at each evet time of the first
phylogeny}

\item{event_times2}{event times of the second phylogeny}

\item{species_number2}{the number of species at each evet time of the second
phylogeny}

\item{distance_method}{(string) absolute, or squared distance?}
}
\description{
Calculates the exact difference between the nLTT
curves of the event times. This includes extinction events.
}
\examples{

# Generate data
n <- 10
b_times_n <- (seq(1, n) / n)
lineages_n <- b_times_n
b_times2_n <- b_times_n * b_times_n
lineages2_n <- b_times2_n

# Calculate nLTT
out <- nLTT::nltt_diff_exact_calc_extinct(
  event_times = b_times_n,
  species_number  = lineages_n,
  event_times2 = b_times2_n,
  species_number2 = lineages2_n,
  distance_method = "abs"
)
#'
}
\author{
Thijs Janzen and Richel Bilderbeek and Pedro Neves
}
