% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nLTTDiff.R
\name{nltt_diff_exact_from_brts}
\alias{nltt_diff_exact_from_brts}
\title{Calculates the exact difference between the lineage through time
curves of two trees' branching times (normalized in time and
for the number of lineages)}
\usage{
nltt_diff_exact_from_brts(b_times_N, b_times2_N)
}
\arguments{
\item{b_times_N}{branching times}

\item{b_times2_N}{branching times}

\item{distance_method}{(string) absolute, or squared distance?}

\item{ignore_stem}{(logical) Should the phylogeny its stem be ignored?}
}
\value{
(scalar) normalized Lineage-Through-Time difference between tree1 & tree2
}
\description{
Calculates the exact difference between the lineage through time
curves of two trees' branching times (normalized in time and
for the number of lineages)
}
\author{
Thijs Janzen
}

