% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nLTTDiff.R
\name{nltt_diff_exact_norm_brts}
\alias{nltt_diff_exact_norm_brts}
\title{Calculates the exact difference between the nLTT
curves of the branching times}
\usage{
nltt_diff_exact_norm_brts(b_times_N, lineages_N, b_times2_N, lineages2_N,
  distance_method)
}
\arguments{
\item{b_times_N}{branching times of the first phylogeny}

\item{lineages_N}{the number of lineages, usually one to the number of lineages}

\item{b_times2_N}{branching times of the first phylogeny}

\item{lineages2_N}{the number of lineages, usually one to the number of lineages}

\item{distance_method}{(string) absolute, or squared distance?}
}
\description{
Calculates the exact difference between the nLTT
curves of the branching times
}
\author{
Thijs Janzen and Richel Bilderbeek
}

