% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nltt_diff_exact_notnorm.R
\name{nltt_diff_exact_notnorm}
\alias{nltt_diff_exact_notnorm}
\title{Calculates the exact difference between the nLTT
curves of the event times. This includes extinction events.}
\usage{
nltt_diff_exact_notnorm(event_times, species_number, event_times2,
  species_number2, distance_method = "abs", time_unit = "since")
}
\arguments{
\item{event_times}{event times of the first phylogeny}

\item{species_number}{the number of species at each evet time of the first
phylogeny}

\item{event_times2}{event times of the second phylogeny}

\item{species_number2}{the number of species at each evet time of the second
phylogeny}

\item{distance_method}{how the difference between the two nLTTs is summed
\itemize{
 \item{"abs: "}{the absolute distance between the two nLTTs is summed}
 \item{"squ: "}{the squared distance between the two nLTTs is summed}
}}

\item{time_unit}{the time unit of the branching times
\itemize{
 \item{"ago: "}{the branching times are postive,
   as these are in time units ago}
 \item{"since: "}{the branching times are negative,
   as these are in time units since present}
}}
}
\description{
Takes branching times such as (for example) as returned by a
\code{\link[DAISIE]{DAISIE_sim}} simulation.
}
\author{
Thijs Janzen and Richel Bilderbeek and Pedro Neves
}
