% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nLTTDiff.R
\name{nltts_diff}
\alias{nltts_diff}
\title{Calculates the exact difference between the lineage through time
curves of a specified tree and a collection of trees}
\usage{
nltts_diff(tree, trees, distance_method = "abs", ignore_stem = TRUE)
}
\arguments{
\item{tree}{A phylogenetic tree}

\item{trees}{One or more phylogenetic trees}

\item{distance_method}{(string) absolute, or squared distance?}

\item{ignore_stem}{(logical) Should the phylogeny its stem be ignored?}
}
\description{
Calculates the exact difference between the lineage through time
curves of a specified tree and a collection of trees
}
\author{
Richel J.C. Bilderbeek
}
