% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_phylogenies.R
\name{check_phylogenies}
\alias{check_phylogenies}
\title{Check if the input is a valid collection of one or more phylogenies}
\usage{
check_phylogenies(phylogenies)
}
\arguments{
\item{phylogenies}{a collection of one or more phylogenies,
where the phylogenies are of type \link[ape]{phylo}.
This collection can both be a list of \link[ape]{phylo} or
a \link[ape]{multiPhylo}.}
}
\description{
Will \link{stop} if not
}
