% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nLTTDiff.R
\name{nltt_diff_exact_brts}
\alias{nltt_diff_exact_brts}
\title{Calculates the exact difference between the nLTT
curves of the branching times}
\usage{
nltt_diff_exact_brts(b_times, lineages, b_times2, lineages2, distance_method)
}
\arguments{
\item{b_times}{branching times of the first phylogeny}

\item{lineages}{the number of lineages, usually one to the number of lineages}

\item{distance_method}{(string) absolute, or squared distance?}

\item{b_times_2}{branching times of the first phylogeny}

\item{lineages_2}{the number of lineages, usually one to the number of lineages}
}
\description{
Calculates the exact difference between the nLTT
curves of the branching times
}
\author{
Thijs Janzen
}

