% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nLTTDiff.R
\name{nltt_diff_exact_norm_brts}
\alias{nltt_diff_exact_norm_brts}
\title{Calculates the exact difference between the nLTT
curves of the branching times}
\usage{
nltt_diff_exact_norm_brts(
  b_times_n,
  lineages_n,
  b_times2_n,
  lineages2_n,
  distance_method
)
}
\arguments{
\item{b_times_n}{branching times of the first phylogeny}

\item{lineages_n}{the number of lineages,
usually one to the number of lineages}

\item{b_times2_n}{branching times of the first phylogeny}

\item{lineages2_n}{the number of lineages,
usually one to the number of lineages}

\item{distance_method}{(string) absolute, or squared distance?}
}
\description{
Calculates the exact difference between the nLTT
curves of the branching times
}
\author{
Thijs Janzen and Richel Bilderbeek
}
