% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nLTT.R
\name{nltt_lines}
\alias{nltt_lines}
\title{Normalized version of the ape function ltt.lines.}
\usage{
nltt_lines(phy, ...)
}
\arguments{
\item{phy}{an object of class \code{"phylo"}}

\item{\dots}{further graphical arguments that can be passed to \code{lines()}}
}
\description{
This is a modified version of the \code{ape} function ltt.lines: add the normalized Lineage-Through-Time statistic of a phylogenetic tree to an already existing plot
}
\examples{
  data(exampleTrees)
  nltt_plot(exampleTrees[[1]])
  nltt_lines(exampleTrees[[2]], lty=2)
}
\author{
Thijs Janzen
}

